Minefield
The rules of the game, according to Wikipedia, are:

The playing area consists of a rectangular field made up of squares. Each square can be revealed when selected, and if the selected square contains a mine, the game is over. If, on the other hand, the square does not contain a mine, one of two things can happen:

A number appears, indicating the number of adjacent squares that contain mines;
No numbers appear. In this case, the game automatically reveals the squares that are adjacent to the empty square, as they cannot contain mines;
The game is won when all squares that do not have mines are revealed.